A connect/postgres integration
This is a node v8.12.0, streaming style postgres 9.6/urban airship connect 1.0.0
integration demonstrating how a person might warehouse connect data for
their app. This is emphatically NOT how Urban Airship does analytics as
part of its Insight analytics service.

It's meant to be illustrative of concerns such as offset tracking,
duplicate safe inserts (via postgres' ON CONFLICT DO NOTHING), and one
example of the analytics a person might do with the resulting data.